Changed question treemap type from short to integer

+++ core/src/com/hackheroes/game/Screens/GameScreen.java
@@ -5,6 +5,7 @@ import com.hackheroes.game.Components.Question;
 import com.hackheroes.game.Components.StatusBar;
 import com.hackheroes.game.MainClass;
 import com.hackheroes.game.Scenes.IndicatorsInfo;
+
 import java.util.Map;
 
 public class GameScreen extends AbstractScreen {
@@ -31,7 +32,7 @@ public class GameScreen extends AbstractScreen {
 
         private int x, y, width, height;
         private String label;
-        private Map<String, Short> effects;
+        private Map<String, Integer> effects;
 
         private AnswerField(int x, int y, int width, int height, String label) {
             this.x = x;
@@ -41,7 +42,7 @@ public class GameScreen extends AbstractScreen {
             this.label = label;
         }
 
-        void render(Map<String, Short> effects) {
+        void render(Map<String, Integer> effects) {
             this.effects = effects;
 
             game.gameShapeRenderer.setProjectionMatrix(game.gameCamera.combined);
